# backend/service.core/docs/swagger.yaml
basePath: /api/v1
definitions:
  categories.ItemResponse:
    properties:
      item:
        $ref: '#/definitions/entity.MenuItemCategory'
        type: object
    type: object
  categories.ListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/entity.MenuItemCategory'
        type: array
    type: object
...
  entity.MenuItem:
    properties:
      allergy:
        type: string
      category_id:
        type: string
      description:
        type: string
      file:
        type: string
      id:
        type: string
      is_menu:
        type: boolean
      is_special:
        type: boolean
      name:
        type: string
      price:
        type: number
    type: object
  entity.MenuItemCategory:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  entity.MenuItemResult:
    properties:
      allergy:
        type: string
      category_id:
        type: string
      category_name:
        type: string
      description:
        type: string
      file:
        type: string
      id:
        type: string
      is_menu:
        type: boolean
      is_special:
        type: boolean
      name:
        type: string
      price:
        type: number
    type: object
  entity.Order:
    properties:
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/entity.OrderItem'
        type: array
      restaurant_id:
        type: string
      status:
        type: string
      table_num:
        type: integer
      total_cost:
        type: number
      user_id:
        type: string
    type: object
  entity.OrderItem:
    properties:
      item_id:
        type: string
      quantity:
        type: integer
    type: object
...
  errors.ErrorResponse:
    properties:
      message:
        example: error message
        type: string
    type: object
  menu_items.CreateMenuItemRequest:
    properties:
      allergy:
        type: string
      category_id:
        type: string
      description:
        type: string
      file:
        type: string
      is_menu:
        type: boolean
      is_special:
        type: boolean
      name:
        type: string
      price:
        type: number
    type: object
  menu_items.GetItemResponse:
    properties:
      item:
        $ref: '#/definitions/entity.MenuItemResult'
        type: object
    type: object
  menu_items.GetListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/entity.MenuItemResult'
        type: array
    type: object
  menu_items.ItemResponse:
    properties:
      item:
        $ref: '#/definitions/entity.MenuItem'
        type: object
    type: object
  menu_items.UpdateMenuItemRequest:
    $ref: '#/definitions/menu_items.CreateMenuItemRequest'
  orders.CreateOrderItemRequest:
    properties:
      item_id:
        type: string
      order_id:
        type: string
      quantity:
        type: integer
      restaurant_id:
        type: string
    type: object
...
  orders.ItemResponse:
    properties:
      item:
        $ref: '#/definitions/entity.Order'
        type: object
    type: object
...
  orders.UpdateOrderItemRequest:
    $ref: '#/definitions/orders.CreateOrderItemRequest'
...
host: localhost:5000
info:
  contact: {}
  description: Restaurants and customers and other great things!
  license: {}
  title: COMP3900 - JAMAR
  version: "0.1"
paths:
...
  /orders/{ord_id}/items/{itm_id}:
    delete:
      parameters:
      - description: The id of the order
        in: path
        name: ord_id
        required: true
        type: string
      - description: The id of the item in the order
        in: path
        name: itm_id
        required: true
        type: string
      - description: The order item's details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/orders.UpdateOrderItemRequest'
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/orders.ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Deletes an order item from an order
      tags:
      - Orders
...
  /restaurants/{res_id}/categories:
    get:
      parameters:
      - description: The id of the restaurant
        in: path
        name: res_id
        required: true
        type: string
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/categories.ListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Gets all menu item categories associated to a restaurant
      tags:
      - MenuItems
    post:
      parameters:
      - description: The id of the restaurant
        in: path
        name: res_id
        required: true
        type: string
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/categories.ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Creates a new menu item category for a restaurant
      tags:
      - MenuItems
  /restaurants/{res_id}/categories/{cat_id}:
    get:
      parameters:
      - description: The id of the restaurant
        in: path
        name: res_id
        required: true
        type: string
      - description: The id of the category
        in: path
        name: cat_id
        required: true
        type: string
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/categories.ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Returns a category from an id
      tags:
      - MenuItems
  /restaurants/{res_id}/menu:
    get:
      description: Returns this restaurant's current displayed menu. It is used by customers only
      parameters:
      - description: The id of the restaurant
        in: path
        name: res_id
        required: true
        type: string
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/menu_items.GetListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Gets a restaurant's current visiable menu items
      tags:
      - MenuItems
  /restaurants/{res_id}/menu_items:
    get:
      parameters:
      - description: The id of the restaurant
        in: path
        name: res_id
        required: true
        type: string
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/menu_items.GetListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Gets all menu items associated to a restaurant
      tags:
      - MenuItems
    post:
      parameters:
      - description: The id of the restaurant
        in: path
        name: res_id
        required: true
        type: string
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/menu_items.ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Creates a new menu item for a restaurant
      tags:
      - MenuItems
  /restaurants/{res_id}/menu_items/{itm_id}:
    get:
      parameters:
      - description: The id of the restaurant
        in: path
        name: res_id
        required: true
        type: string
      - description: The id of the item
        in: path
        name: itm_id
        required: true
        type: string
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/menu_items.GetItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get a menu item's details
      tags:
      - MenuItems
    put:
      parameters:
      - description: The id of the restaurant
        in: path
        name: res_id
        required: true
        type: string
      - description: The id of the item
        in: path
        name: itm_id
        required: true
        type: string
      - description: The menu_item's updated details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/menu_items.UpdateMenuItemRequest'
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/menu_items.ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Updates a menu item for a restaurant
      tags:
      - MenuItems
...
swagger: "2.0"